Add ManyToMany tagging
Adds the automaton parts for the new many to many tagging.

We define a new constant AutomatonRelation.MANY_TO_MANY (="ManyToMany")
which we use as signal for many to many relations since we can't
determine that by the model alone in all cases.

To have a many-to-many relation that can be edited with the
<AssociationSelector/> and that works with the WorkingSet, we now need
to tag both relations of link type with AutomatonRelation.MANY_TO_MANY.

Validates on startup that such link type has exactly two such relations.