Add AutomatonDomain
Adds AutomatonDomain, our own DomainQL builder class.

It contains a convenience method that shortens the necessary
GraphQLConfiguration in the individual applications by
providing an initial DomainQLBuilder instance that already has the meta
data generators registed and that already contains the necessary scalar
types we expect an automaton application to have.