Introduce unified JSON output / error handling
One of the most common error scenarios is that a user is working within
an automaton app and his session either expires time-based or the server
restarts or whatever.

Now the user resorts a data grid, executes a new process. Most of the
time this will happen via AJAX/fetch request with a application/json
content type.

By default Spring Boot nevertheless just answers to the now
unauthenticated request with the login HTML view, which of course isn't
ok with the JSON parser.

This commit contains the library-side part of the solution for now.

The whole solution goes as follows. In the configure method of our
SecurityConfiguration we add the following builder calls:

-------------------------------- snip ------------------------------

    .addFilterBefore(concurrentSessionFilter(), ConcurrentSessionFilter.class)
    .sessionManagement()
    .sessionAuthenticationStrategy(sessionAuthenticationStrategy())
    // configure the number of concurrent sessions, -1 for unlimited
    .maximumSessions(-1);

    // ... definitions ....

    @Bean
    public SessionRegistry sessionRegistry()
    {
        return new SessionRegistryImpl();
    }

    @Bean
    public SessionAuthenticationStrategy sessionAuthenticationStrategy()
    {
        return new ConcurrentSessionControlAuthenticationStrategy(
            sessionRegistry()
        );
    }

    @Bean
    public Filter concurrentSessionFilter()
    {
        return new ConcurrentSessionFilter(
            sessionRegistry(),
            new AutomatonSessionExpiredStrategy()
        );
    }

-------------------------------- snap ------------------------------

We use ConcurrentSessionControlAuthenticationStrategy which as a
side-effect registers a ConcurrentSessionFilter in the spring security
filter chain.

Now the default implementation of that still doesn't work us, as it
sends plaintext errors.

But the ConcurrentSessionFilter allows the registration of a
SessionInformationExpiredStrategy which we can use to fix things.

This commit implements a SessionInformationExpiredStrategy that checks
the content type and sends a JSON error response.

Unified Controller Output
-------------------------

The easiest way to handle this scenario and other kinds of cross-cutting
scenarios is to standardize the JSON response so that we can intercept
JSON requests and answer them without having to deal with a lot of
different URIs and formats.

Volume-wise, we already do 90% of our requests with GraphQL which offers
a solution for exactly the problem we have now.

So we just use the GraphQL convetions for data responses, whether we use
GraphQL or not.

All payload data is available under a "data" key within the response.
For errors there is also an "errors" key. The error key might be present
even though the data key is also present. If there are no errors, the
"errors" key should *not* be present. If the data acquisition fails, the
"data" key should not be present.

The errors array contains error object with a "message" key containing
the error message and a "path" key containing a GraphQL path as string
array. For our purposes, the path will be almost always be an empty
array. GraphQL also offers a location key that points to the error
location within the GraphQL query, which is also not really applicable
here.

This commit changes the data format for all framework controllers
(ProcessController, AttachmentController, ScopeSyncController)

See the corresponding commit in automaton-js for the client-side of
these changes.