Update to JDK 17
Update to JDK 17 and dependencies.

Most of it is just javax.* to jakarta.* changes.

The update to the newest graphql-java comes with some unfortunate
changes.

So far, the GraphQL environment reported only the properties we actually
set in the query parameters. E.g. for a InteractiveQuery based request,
the incoming QueryConfigInput might be just have a condition parameter
and nothing else.

The input conversion would then instantiate the Java object and only
override the condition property with a new ConditionScalar object.

Now we are receiving a value for every single property even if it wasn't
set in the original POSTed JSON blob.

So where you could just have a property field "private int pageSize = 10;"
before and that would stay that way you now have to define a formal default
value on the getter.

     @GraphQLField(defaultValue = "10")
     public int getPageSize()
     {
         return pageSize;
     }

Also note that the logic now tries to write a "null" value for every
property, regardless of whether it is