Server-side impl of GraphQL data exporting
Adds the GraphQLExport controller with is an alternate GraphQL service
for GraphQL data exporting. It serves whatever content type the
GraphQLExporter requires.

The client-side chooses one of the

    de.quinscape.automaton.runtime.export.GraphQLExporter

spring beans defined in the application. This exporter receives the
output of the GraphQL execution and can produce the desired content type.

Adds JSONExporter as simple example.