Implement ExcelExporter
Adds a customizable GraphQLExport implementation for Excel files.

You can define as many excel exporters as you need in your service
configuration.

For example

    @Bean
    public ExcelExporter excelExporter(@Lazy DomainQL domainQL)
    {
        return ExcelExporter.newExporter(domainQL)
            .withExportStrategy(
                // default strategy with custom names
                new ExcelExporter.OneQueryStrategy("Automaton-Test Export", "automaton-test-$now.xlsx")
            )
            .build();
    }

defines the default OneQueryStrategy with custom names for the Excel
sheet and the export file name. If you don't define a strategy, you get
the same strategy with default names.

ExportStrategy is a simple strategy interface that controls the export
of GraphQL queries to one or multiple Excel sheets.

Here an example for a two-sheet export:

    @Bean
    public ExcelExporter bazExporter(@Lazy DomainQL domainQL)
    {
        return ExcelExporter.newExporter(domainQL)
            .withExportStrategy((ctx, queryContext) -> {

                final GraphQLQueryContext.MethodResult bazes = queryContext.getMethodResult("iQueryBaz");
                final GraphQLQueryContext.MethodResult appUsers = queryContext.getMethodResult("iQueryAppUser");

                // note how you also receive a schema reference for the query data, so we could also name sheets by
                // domain type etc pp
                log.info("Method ref type: {}", bazes.methodReference().toString());

                ctx.addSheet("Baz Report", bazes);
                ctx.addSheet("AppUser Reference", appUsers);

                return "baz-report-$now.xlsx";
            })
            .build();
    }

The strategy interface allows you to define as many Excel exports you
need with the titles and pages that make sense for your domain.

Adds POI as optionally dependency. If you want it, declare it.