JOOQTransformers: Enable shortcutting
The way we currently do query-executions is still quite funky and should
be revisited soon. Both to reduce this kind of mess but also to improve
many-to-many join support.

What happens now is that conditions for queries are transformed
multiple times where field references to fields outside of the current
query execution are turned to null values in between which are then
optimized away afterwards so that only the field conditions relevant to
the current query remain.

This wasn't correctly accounted for in both newish JOOQTransformers,
i.e. the parts of our FilterDSL that do not map to reflective
invocations on the field objects but are implemented as auxiliary
system.

These are concat() where the FieldDSL method did not work out due to
method signature mismatches and the toString() method, a specialized
cast operation (where the problem was that we have no way to reference
the classes needed for that cast operation in our FilterDSL)

This commit changes both transformers to gracefully handle intermediary
null values.