FORK: JasperParameterFunction: Add Type-Support
Adds Type-Support when building SQL-Output from JasperParameterFunction.
To do so the Jooq Field from the left side of the Condition was passed
down into the JasperParameterFunction. The JasperParameterFunction reads
the DataType of the Jpooq-Field and generates an appropriate SQL Output
to use in JasperReports.

This commit restores some improvised code for an upcoming release. The
method on the main branch does not really work and will be reworked in
time.

This branch switches to its own namespaced maven artifact.