Refactor Filtertransformer
The computed value node in our FilterDSL is quickly becoming the main
way of plugging in custom project specific stuff into the
InteractiveQuery mechanism. (We should think about removing the filter
context stuff which is not used as this point).

As such it is important that we know a little bit more about the
invocation context of our custom code, most importantly the type context
of the surrounding expressions.

This commit changes the signature of ComputedValue

public interface ComputedValue
{
    Object evaluate(ComputedValueContext ctx);
}

The new context encapsulates the former name and args parameter and also
contains a type context which reflects what we know about the types
involved.

If it is a simple field reference, we have the type and field of that
reference available.

Otherwise we have at least the scalar type including decimal precision
to fully support BigDecimal.

The refactoring got a little more complex in that we now have a
dependecy between the FilterTransformer and our JOOQ instance, but I
don't think that's a problem and was only caused by us using the
statically generated JOOQ knowledge so far.

As the tests have shown, where you need the filtertransformer you need a
JOOQ instance anyway to do anything meaningful.