FilterDSLScalar: Refactor toString()
The JSONification of the condition tree can fail when it contains e.g.
Date or Timestamp values which are not jsonable per se.

It would be nice to have the decompileFilter.js we have on the client
side, that would make for a nice concise output.

For now we change it to a simple toString() to get rid of the error.