ExcelExporter: Improve styling
First of all we need to tell POI that our values we set with
.setCellValue(LocalDateTime) and .setCellValue(LocalDate) are actually
date fields and we don't want to have it displayed as "40908". I thought
that was clear.

We take two default styles for Date and DateTime fields.

Adds a second pass that goes over all columns and calls
autoSizeColumn(int) for it.